Moved the scoreboard to the index page for the site.
Added 'user detail' and 'problem detail' to the urls for judge
application

On problem detail page, created the ability to 'regrade' or 'delete' a
user's result

minor fix in Problem.possibleScore(user) to never return a score that is
negative.

Fixed CSS/HTML issues regarding tbody & thead on Scoreboard.

Made login not required to view the scoreboard page.

[BUG FIX] Users were dislayed in wrong order on scoreboard. Reverse the
listing so it goes from highest --> lowest, not low --> high.